+++ cmd/flags.go
@@ -5,6 +5,7 @@ import (
 	"log/slog"
 	"os"
 
+	"github.com/gnames/dwca/internal/ent"
 	dwca "github.com/gnames/dwca/pkg"
 	"github.com/gnames/dwca/pkg/config"
 	"github.com/spf13/cobra"
@@ -29,6 +30,23 @@ func rootDirFlag(cmd *cobra.Command) {
 	}
 }
 
+func fieldsNumFlag(cmd *cobra.Command) {
+	s, _ := cmd.Flags().GetString("wrong-fields-num")
+	switch s {
+	case "":
+		return
+	case "stop":
+		opts = append(opts, config.OptWrongFieldsNum(ent.ErrorBadRow))
+	case "ignore":
+		opts = append(opts, config.OptWrongFieldsNum(ent.SkipBadRow))
+	case "process":
+		opts = append(opts, config.OptWrongFieldsNum(ent.ProcessBadRow))
+	default:
+		slog.Warn("Unknown setting for wrong-fields-num, keeping default",
+			"setting", s)
+		slog.Info("Supported values are: 'stop' (default), 'ignore', 'process'")
+	}
+}
 func archiveFlag(cmd *cobra.Command) {
 	archive, _ := cmd.Flags().GetString("archive-format")
 	if archive != "" {