+++ pkg/dwca_test.go
@@ -5,6 +5,7 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/gnames/dwca/internal/ent"
 	"github.com/gnames/dwca/internal/ent/dcfile"
 	"github.com/gnames/dwca/internal/ent/diagn"
 	dwca "github.com/gnames/dwca/pkg"
@@ -58,7 +59,7 @@ func TestCoreData(t *testing.T) {
 	}
 	for _, v := range tests {
 		path := filepath.Join("testdata", v.file)
-		cfg := config.New(config.OptWithSloppyCSV(true))
+		cfg := config.New(config.OptWrongFieldsNum(ent.ProcessBadRow))
 		arc, err := dwca.Factory(path, cfg)
 		assert.Nil(err)
 		assert.Implements((*dwca.Archive)(nil), arc)
@@ -90,7 +91,7 @@ func TestCoreStream(t *testing.T) {
 	}
 	for _, v := range tests {
 		path := filepath.Join("testdata", v.file)
-		cfg := config.New(config.OptWithSloppyCSV(true))
+		cfg := config.New(config.OptWrongFieldsNum(ent.ProcessBadRow))
 		arc, err := dwca.Factory(path, cfg)
 		assert.Nil(err)
 		assert.Implements((*dwca.Archive)(nil), arc)
@@ -136,7 +137,7 @@ func TestExtensionData(t *testing.T) {
 
 	for _, v := range tests {
 		path := filepath.Join("testdata", v.file)
-		cfg := config.New(config.OptWithSloppyCSV(true))
+		cfg := config.New(config.OptWrongFieldsNum(ent.ProcessBadRow))
 		arc, err := dwca.Factory(path, cfg)
 		assert.Nil(err)
 		assert.Implements((*dwca.Archive)(nil), arc)
@@ -172,7 +173,7 @@ func TestExtensionStream(t *testing.T) {
 	ctx := context.Background()
 	for _, v := range tests {
 		path := filepath.Join("testdata", v.file)
-		cfg := config.New(config.OptWithSloppyCSV(true))
+		cfg := config.New(config.OptWrongFieldsNum(ent.ProcessBadRow))
 		arc, err := dwca.Factory(path, cfg)
 		assert.Nil(err)
 		assert.Implements((*dwca.Archive)(nil), arc)
@@ -216,7 +217,7 @@ func TestSynDiagnose(t *testing.T) {
 
 	for _, v := range tests {
 		path := filepath.Join("testdata", "diagn", "synonyms", v.file)
-		cfg := config.New(config.OptWithSloppyCSV(true))
+		cfg := config.New(config.OptWrongFieldsNum(ent.ProcessBadRow))
 		arc, err := dwca.Factory(path, cfg)
 		assert.Nil(err)
 		assert.Implements((*dwca.Archive)(nil), arc)